+++ Editor/EditorWindow/SWQuickAssetPaletteWindow.cs
@@ -48,7 +48,7 @@ namespace SWTools
         /// <summary>
         /// Quick Asset Palette 창을 엽니다.
         /// </summary>
-        [MenuItem("SWTools/Quick Asset Palette")]
+        [MenuItem("SWTools/Utils/Quick Asset Palette")]
         public static void ShowWindow()
         {
             SWQuickAssetPaletteWindow window = GetWindow<SWQuickAssetPaletteWindow>();