+++ Editor/EditorWindow/SWTestToolsWindow.cs
@@ -124,7 +124,7 @@ namespace SWTools
         /// <summary>
         /// Test Tools 창을 엽니다.
         /// </summary>
-        [MenuItem("SWTools/Test Tools Window %#t")]
+        [MenuItem("SWTools/Debug/Test Tools Window %#t")]
         public static void ShowWindow()
         {
             SWTestToolsWindow window = GetWindow<SWTestToolsWindow>();