+++ Runtime/Attribute/SWConditionAttribute.cs
@@ -1,11 +1,6 @@
 using UnityEngine;
 using System;
 
-
-#if UNITY_EDITOR
-using UnityEditor;
-#endif // UNITY_EDITOR
-
 namespace SWTools
 {
     /// <summary>